+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_params.py
@@ -21,7 +21,6 @@ from azure.cli.core.commands.parameters import (
     resource_group_name_type,
     get_one_of_subscription_locations,
     get_resource_name_completion_list)
-from azure.mgmt.compute.containerservice.models import ContainerServiceOrchestratorTypes
 from azure.cli.command_modules.acs._validators import validate_create_parameters, validate_ssh_key, validate_list_of_integers
 
 
@@ -60,6 +59,8 @@ def _get_default_install_location(exe_name):
 
 name_arg_type = CliArgumentType(options_list=('--name', '-n'), metavar='NAME')
 
+orchestratorTypes = ["Custom", "DCOS", "Kubernetes", "Swarm", "DockerCE"]
+
 storageProfileTypes = ["StorageAccount", "ManagedDisks"]
 
 register_cli_argument('acs', 'tags', tags_type)
@@ -70,7 +71,7 @@ register_cli_argument('acs', 'name', arg_type=name_arg_type, configured_default=
 
 register_cli_argument('acs', 'resource_group', arg_type=resource_group_name_type)
 
-register_cli_argument('acs', 'orchestrator_type', options_list=('--orchestrator-type', '-t'), **enum_choice_list(ContainerServiceOrchestratorTypes))
+register_cli_argument('acs', 'orchestrator_type', options_list=('--orchestrator-type', '-t'), help='DockerCE only available since api version 2017-07-01', **enum_choice_list(orchestratorTypes))
 # some admin names are prohibited in acs, such as root, admin, etc. Because we have no control on the orchestrators, so default to a safe name.
 register_cli_argument('acs', 'admin_username', options_list=('--admin-username',), default='azureuser', required=False)
 register_cli_argument('acs', 'api_version', options_list=('--api-version',), required=False, help='Use API version of ACS to perform az acs operations. Available options: 2017-01-31, 2017-07-01. (2017-07-01 in preview, only in ukwest and uksouth). Default to use the latest version for the location')